feat: add new query

+++ src/main/java/service/GenericDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface GenericDao<T> {
     List<T> findAll();
     T update(T object);
-    T get(Long id);
+    T get(Integer id);
     void delete(T object);
     void insert(T object);
     boolean exists(Long id);
